pyproject.toml
[project]
name = "autolog"
version = "1.2.0"
description = "automates the process of logging work entries into Jira"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "customtkinter>=5.2.2",
    "jira>=3.8.0",
    "keyring>=25.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "commitizen>=4.8.2",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pyinstaller>=6.13.0",
    "ruff>=0.11.8",
]

[tool.commitizen]
name = "cz_conventional_commits"
update_changelog_on_bump = true
version = "1.2.0"
version_files = [
    "pyproject.toml:version",
    "__version__.py"
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = ["autolog"]
skip = [".venv", "build", "dist"]

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
unsafe-fixes = false 
preview = true

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "N",  # pep8-naming
  "B",  # flake8-bugbear
]

ignore = [
  "E203",  # Whitespace before ':' (conflicts with Black)
]

exclude = [
  ".git",
  ".venv",
  "build",
  "dist",
  "__pycache__",
]

[tool.ruff.lint.isort]
known-first-party = ["autolog"]
combine-as-imports = true
force-single-line = false


[tool.pyinstaller]
onefile = true
console = false
name = "AutoLog"
add-data = ["autolog/*:autolog"]
